Only validate one declaration section at a time
There was a bug where all list-type questions (checkboxes) would have their value set to None when any page of the declaration was saved that did not include that question.  This was because the submitted form data was checked against the entire declaration content rather than just the section that has been submitted.