Pass question_advice into template explicitly
Before, we were setting `question_advice` as a global template
variable, which meant that the question template had access to
it, but so would anything else in the template.
Now, we're passing it into the question template explicitly.
I thought it was a bit neater the first way, but it's probably
more correct this way.