Make footer categories and meta grid rows
This includes styles for containing floats so
doesn't need the `overflow: hidden` hack.

Also removes the side-padding from the categories
wrapper as its containing element now has its
width defined by the %site-wide-container block
(brought in with the new govuk template) with sets
it to 960px on desktop so now needs to no spacing
on either side;

Also moves the padding-bottom for the categories
wrapper onto it's child columns so they push away
from the `<hr>` below.