Update links in header
Remove "Create supplier account" link from the header
View your account link comes before the Log out link.
Unlike in the header for the buyer app, we don't have a conditional
statement on the "View your account" link because nobody can even
get to the supplier app without being logged in with the correct
role. So it always sends logged-in users to the supplier dashboard.

We changed from current_user.email_address to
current_user.is_authenticated() to check if a user is logged in or
not.