Built first page of contract signing flow + tests
Recycling %summary-item-heading from toolkit for page-subheading
class.

Test that we can load the page and that the correct lots show up.
Also explanded the `framework` test helper function to return all
of the G-Cloud lots, as well as added a `get_g_cloud_8` test
fixture.

Link directly to the PDF of the framework agreement contract
to reduce the (potentially) jarring change of context.