Display the default evaluation type for brief responded to
The default evaluation types defined in the frameworks content are not stored in the brief object JSON at all, and are only added to the brief when `.summary` is called in the content loader.
So in order to be able to display all evaluation methods that will be used, including the defaults, we need to load the `edit_brief` manifest and call `.summary` with the current brief.