Refactor signer_details routes + fix tests
Many of the setup steps are the same for the GET and POST routes
for the signer details page. (And, in the case of a failing POST
request, so are the steps before rendering the template.)

Merged both the GET and POST route together to avoid some to
avoid some repetition and then fixed a few breaking tests.