Use organisationName from declaration
Suppliers need to enter a name for their company when they sign
up to the Digital Marketplace, but it's only in the declaration
that we ask then what they should legally be referred to in
legal documents.
Thus, we should be using the legal name they provided in their
declaration throughout the contract submission process rather
than the name they entered when they signed up.