Show uploaded time if filename not in session
The session variable is set when a document is uploaded initially,
but a supplier could come back to complete their submission after
the session has expired, in which case we would only have the
uploaded time.
Fall back to just showing the uploaded time if we don't have the
filename.

Add tests + mock s3 where necessary.