Editing signer details redirects to contract review page
For the user journey where they are on the contract review page and
choose to edit their signer details, if their input is successful we
want them to be redirected back to the contract review page rather
than progressing through the usual flow and going to the upload
page.

We signify this by if they have a filename in the session.