If file already uploaded, allow no file chosen on signature-upload
If a file has already been uploaded to s3 and the user visits the
signature-upload form then we will allow them to 'Save and continue'
without having chosen a new file to be uploaded. This means you can
review what file has been uploaded previously (communicated on the
page to the user) and if you do not want to change it, you can
continue to use it.