Framework dashboard changes after submitting contract
The `contract_submitted.html` page will replace the regular
framwork dashboard page after signing and uploading a
contract.
It should contain the name of the person who signed the
contract, information about the user who actually submitted
the file and the time it was uploaded.

Note that, currectly, this page won't change when CCS
countersign an agreement.  That's another story.

Add a test to make sure the information we expect is on
the page.