framework dashboard: fix pluralization of lot units
by rehashing the way we pass lot data to templates. this code seems to have missed out on the addition of unitSingular and unitPlural
to the lots table, so use that. rather than doing laborious cherry-picking of lot dict items we may as well just pass copies of the
api-supplied lot dicts themselves with our added data annotated on.