Send a virtual pageview on the dashboard
If the flag we set in the flash is present we set
a data attribute in the HTML which the analytics
JS gets the data from to send a virtual pageview.

This required us to add some filtering of the
flash messages we use for the notification
banners. Previously they were the only messages
sent into the page so this wasn't needed.