Set custom dimension for supplier eligibility based on template
For the create brief response url, we will inspect the html content
of the page to decide if we should set a custom dimension, and if
so, what it's value should be.

A `data-reason` attribute in the html represents a supplier not being
eligible for a brief. We will send it's value.

No `data-reason` attribute in the html represents a supplier being
eligible for the brief.