Redirect to contract variation after returning framework agreement
If a user returns their framework agreement but has not agreed
to the framework variation then redirect them to the variation.

If a user returns their framework agreement and has agreed to the
framework variation then redirect them to the framework dashboard.

If there are no framework agreement variations for a framework then
the user should redirect to the framework dashboard.

Functionality is built under the assumption that a framework will
have no more than one variation. Removing this assumption would
add a reasonable amount of complexity, such as which variation
should be redirected too. At the moment, we have no foresight that
a framework having more than one variation is coming any time soon
so should not be worth investing in.