Remove S3 dependencies from framework routes
Now we have paths stored in FrameworkAgreements for signed and countersigned documents
we no longer need to trawl through S3 to find if these exist.

If an agreement has been signed or countersigned in the database we can instead use
the saved paths to get the file.

Routes updated here:
 * Framework dashboard
 * Download agreement file
 * Framework updates