Remove `get_agreement_document_path` from framework views
When setting filenames for returning agreements we use the new
`generate_timestamped_document_upload_path`.

When finding a file (not is not a returned agreement or countersigned
agreement), we still rely on a expected pattern for our file paths
but we use the slightly more generalised util `get_document_path`.