Revert to generic behaviour for `download_agreement_file` route
We revert the `download_agreement_file` to be dumb and not
know about any specific supplier framework paths such as `agreementPath`.
To do this we use the util `degenerate_document_path_and_return_doc_name`
for individual routes to take a path (such as the `agreementPath`)
and strip it down to it's generic document name which can be passed
to the `download_agreement_file` route.