service submission: force return_to_summary flow if framework.framework == "dos"
in the process adding a handy docstring to document the intended behaviour of
of the url parameter. rename some variables for clarity. no longer submit
the `return_to_overview` action in the other case as it's never actually
consumed by the view and so is just confusing. returning to the overview
can just be considered the default behaviour.