Update FrameworkAgreements directly for pre-G-Cloud-8 frameworks
This updates saving signed agreements to use the FrameworkAgrement routes rather than the
update_supplier_framework route to save agreements.

Rollbacks are no longer needed on error because we don't care if a draft agreement is left
abandoned in the database, a new agreement will be created on subsequent attempts to save
if an error occurs.