Create framework agreement as start of signing process
We have changed the `return your signed signature page` link to
a button that will trigger a new framework agreement object to be
created that can be used in the signing process. This needs to be
created at the beginning so the ID of that agreement can be used in
the url of the signer details page.