Use stricter check on signedAgreementPath to decide redirect location
Instead of checking a file name in the session, we check a
`signedAgreementPath` in the framework agreement as this is a
stricter check and avoids the case were they may upload a file and
then restart the whole process (i.e. create a new draft). This
check allows us to decide if to skip straight to the contract
review page or not.