Add missing |safe declaration for notification banners
Some banners are showing links inside the message, so we need to
mark them as safe since we've removed the `safe` filter from the
template itself.

This feels like a better approach, since the decision is made on
a per-message basis, as opposed to in the template.

This also fixes an XSS in the service submission notification
caused by displaying the service name in a message marked as safe.
Instead, we only mark the message fragment with the link as safe.
Appending a string to the Markup then escapes the added string.