Pass draft brief response to start page
If a draft brief response already exists, it needs to be passed to the
start page template so that the correct url can be used when editing
said draft brief response.

If the brief response is a draft, the template will render a link
button, as by continuing so state is being changed. If their is no draft
brief response then a button is rendered, as a POST route will be hit to
create the response and state will be changed.