Load new edit brief response manifest
This will need to be changed before this work is merged to master.

There is another piece of work which needs doing in the frameworks repo
to move the current edit brief response manifest and questions to legacy
versions to make way for the new flow manifest and questions.

As this hasn't happened yet, we don't know what this manifest is called.
When we do, we'll change this.