Add route to edit brief responses
The new flow for editing brief responses allows the supplier to step
through each page question by question. This new route allows that
functionality. It uses the different sections of the new manifest (still
to be created) and cycles through them, saving the updates to a
brief_response.

If there are no new sections to load in the manifest, the submit brief
response route on the apiclient is called to mark the brief as
submitted.