Check if new brief responses submit ok before viewing
This is a temporary solution, as there is more work to do on updating
the view_response_result. It allows the new style of brief response to
be given the correct result state by checking for the new
`essentialRequirementsMet` key, before checking that all
`essentialRequirements` are true. It allows an end to end experience in
the app, but may not be how we eventually deal with this situation.