Redirect user to correct flow for brief
If a user alters the URL from `/start` to `/create` or visa versa they
could potentially respond to a legacy brief with a new flow response, or
visa versa.

This commit will check the `publishedAt` date of the brief and compare
it to the flag and if the request is for the wrong view, it will
redirect it to the correct view.