Add more detail to XSS assertions
When checking that HTML and markdown sent in has
been escaped it was a bit unclear that finding the
same text as a text node meant it had been
escaped.

This adds some more comments and assertions to
make clear that the tag we intended to inject
isn't added as well as the text being as expected.