Kill the cryptography bug
We've had a problem installing cryptography on local machines,
one solution to which is to make sure your is updated.

This pull request would add a step to the Makefile to always
upgrade pip before installing the rest of the requirements.

If we want to do this, I can port it to the other apps.

[Issue on github](https://github.com/pyca/cryptography/issues/2350#issuecomment-179321979)