Make results page tables work with new flow data
Makes an assumption that boolean list answers
which are hashes (rather than booleans) will
have either an 'evidence' key or, in the absence
of that, will have a 'yesNo' boolean.

Also includes the addition of visible column
headings to both versions of the tables.

Includes limiting results table headers to those
with evidence.