Add new page when starting supplier declaration
This commit adds a new page into the supplier declaration flow when
starting their declaration. This page will only be shown if no
questions have been answered in the declaration, so if they leave
and come back the page will be skipped and they will drop straight
back into questions.