Make results data section work with dynamic lists
This view was originally changed to allow both
of the two types of brief response data format we
now have.

This was based on the assumption that we could use
the same content manifest to display both which is
incorrect.

These changes assume the content (driven by the
right manifest) will match the format of data sent
in. This means questions that are dynamic lists
are identified correctly so can have their own
logic.

One caveat: we had a test for nice-to-have
requirements which made sure a table was not
generated for them if the brief contained none.
This needed to work for when they were both
boolean and dynamic lists (see line 40).