Update results page tests confirmation check
The h1 used to tell users if their application was
sent or not. This is now handled by the
notification banner. The h1 now just says 'Your
application'.

Note: the test for the new flow when users got to
the results page without complying with the
essential requirements was removed. This is no
longer possible in this flow.