Split test_supplier_details_shown into two tests
The suite passed before this change due to the
legacy flow setting its own version of this test,
overwriting the shared one.

I also noticed this test can't be shared because
the assertions rely on frameworks content which
differs between flows for the availability field.

This also contains some refactoring to set a
default brief_responses list on the legacy flow
class to match how it's done on the other one.