Use `frameworkSlug` instead of `frameworkFramework` for `data-reason`
As `frameworkFramework` is changing value, our data-reason will
be changed when a supplier is not on the framework. This in itself
will cause no issues for our analytics however after discussion
with Ash, we can use this as an opportunity to change it to
something more specific to the framework so we can differentiate
between iterations.