Use frameworkFramework when rendering service name
In the list view of services, if a service is a DOS service we don't
provide a link to edit that service. This was achieved by comparing the
services `frameworkName` to the string
`digital-outcomes-and-specialists`. This wouldn't work for DOS2.

We now compare using the recently update `frameworkFramework` which will
be `digital-outcomes-and-specialists` for all DOS services, regardless
of the framework iteration.