Prefill declaration section if allowed
On the edit page for a declaration section this checks if the current section has already been successfully saved (by checking for validation errors for the current section).  If no previous answers exist and the supplier has agreed to reuse of a previous declaration and the section allows prefilling then it fetches the old declaration and pre-populates the answers on the page for review.