Move from integration tests to unit tests if supplier ineligible
Several of our brief routes hand off to
`_render_not_eligible_for_brief_error_page` if the supplier is not
eligible for the brief. We were originally testing this functionality
by integration tests covering some but not all of the routes
which use it. This was not a great solution as it either meant we
were duplicating our tests to cover several different ineligiblity
situations for each route, or meant that we were not properly
testing all the possible situations for each route.

As we are about to remove the 'legacy flow' for suppliers applying
to a brief response, this route had the detailed tests covering
all the situations when a supplier was not eligible for the brief.
As these tests will shortly be removed it left the situation that
we could move these tests to a different route or we could instead
unit test the functionality.

Now we have unit tests for `_render_not_eligible_for_brief_error_page`
and check all routes that use it do hand off to this method when
the supplier is not eligible.

One thing to note is that whilst we will be checking that the
correct variables are handed to the template, we do not have tests
that test the template. We should add functional tests for this
instead.

Also in a few places I have added autospec to our mocks as this
reduces testing errors with incorrect assert statements.