service submission pages: merge generation of "You need to make the supplier declaration" banner, retarget
previously this code was duplicated between the submission and submission lot pages
and linked to the first question page of the declaration (the old way of doing things).
instead we now point at the declaration overview or the "start declaration" page
depending on the status of the declaration. also improve testing coverage of this
banner's presence/absence.