Add counter element to checkbox tree with JavaScript
Previously, we were just overwriting the text of the counter
element with JavaScript.  Now, we want to preserve what was there
before but we want the new stuff on a separate line.

We want the whole thing to stick to the top of the window, so it
all needs to live inside the sticky-header class, but we only
want the counter to update when clicks happen (plus we only want
a counter if javascript is enabled), so know we just append the
counter as part of our initialization.