Change "Pass" to "Successful" and deal with failed lots
Suppliers prefer the word "Successful" to "Pass", and it ma=tches what we say in the emails to them.

Also, turns out the page was not dealing with DOS-style frameworks properly and showing failed lots as "No application".  This fixes it so that if a lot was submitte dbut failed it will show "Unsuccessul" rather than "No application".