Store supplierId in brief question audit event
When a user asks a clarification question on a brief, we want
to be able to be able to easily find the supplier the user
belongs to so that we can notify them of changes to the brief.

We're pretty lax about what goes into audit events, so it seems
like a good idea to keep track of the supplier ID in the audit
event created when a clarification question is asked about a brief.