Do not show table rows for empty questions
For a supplier viewing their published services, if any question
is empty, then we should not show a summary table row for it.
An example of this is DOS briefs, where if the supplier is on
digital specialists lot but only on one specialist role, they will
not see summary table rows for all the other specialist roles.

Note, currently the only questions in our `view_services` manifests
that could be a valid submitted service and be empty are:
- specialist role and specialist price
- outcomes capabilities

Therefore, no pages that already exist (such as viewing GCloud
services) will be affected by this change, only the new pages that
show DOS services.