Preserve line breaks for viewing your brief responses
We're going to allow newline formatting for `textbox_large`
content submitted to briefs, as well as for both submitted
questions and answers for briefs.

Here, we're updating utils to bring in the `preserve_newlines`
filter and then updating the "View your application" page so
that brief response answers are formatted with linebreaks.

**Note**

It's not going into the service editing summary page (either here
or in the admin app) because the way we group the questions in the
service editing screen relies on line breaks to differentiate the
multiquestions from each other. If we allow the
`preserve_line_breaks` filter on the service editing screen,
it becomes very hard to tell how many questions there are and
what you've actually answered to each of them.
(We don't have this problem on the public service listing though,
so we're putting it into the buyer app.)