Update S3ResponseError constructor
Some of our tests were importing an S3ResponseError from
boto which has since been updated in v27 of the utils
library.
The old arguments we used to pass into an S3 error were not
being accepted anymore so I've updated them.

Seems like our test shouldn't be relying on this kind of thing,
but this was a relatively easy fix.