Conditionally show registration information
Rows should only appear if there is data for that field.
Additionally, if a Companies House number is set then the registration country and other registration number fields should be hidden whether there is data for them or not. And the registration number is only shown if the business was not established in the UK.