Bugfix: not validating Companies House strictly enough on frontend
We recently introduced a validator on the API to restrict Companies House numbers to valid formats, but didn't update the frontend flow here at the same time.

Because nothing gets posted to the API until the final step of this flow it is currently possible to submit an invalid CH number here and then have the API post request fail when submitting the final step.

It's not ideal to have the same regex in two places, but this is a current production bug and this will fix it for now.