Send supplier create emails with Notify
As part of the supplier sign up flow a user creation email is sent to a
given email address. This was done by Mandrill, now it's done by Notify.

The functionality to do this for supplier users and buyer users is
really similar so the code to do it has been abstracted to dmutils. We
import it here and use it.