Remove old redirect for user creation
Supplier creation was moved to the user app. Any tokens created before
the move will have linked to the supplier app still. Tokens are valid
for one week, so we needed to maintain the functionality of those
tokens by keeping a redirect on the old route to the user app.

It's been over a week so all tokens will now link to the user app. We
can remove this redirect and associate test.