Create a summary row for unanswered questions
The service definition document field was optional for G-Cloud 9.  We want
a row to be shown for this field even if no file has been uploaded yet.
This adds a summary row with the name of all editable fields, including
those that have not been answered yet.  It shows the "empty_message" for
the question if it exists.