Remove unused styles and imports.
_summary.scss already imported in main application.scss; import here
should be redundant. `lede` and `visual-emphasis` classes not used
in supplier templates.

`page-subheading` extends `summary-item-heading` but doesn't change the
styling applied in any way, so is basically an alias. In aiming to
remove redundant styles, I've removed this and replaced the classes
applied in the templates to use `summary-item-heading`. Although we lose
some semantic meaning, it does mean one less thing to maintain/be aware
of.