Remove unnecessary `with self.app.test_client()` calls in `test_frameworks.py`
These calls are, and always have been as far as I can tell, unnecessary.

I guess they are some sort of copy-paste artefact.

The only changes in this PR are removing the with self.app.test_client() calls and sorting out the indenting

This makes the test more readable and easier to maintain.