Change object permissions to 'bucket-owner-full-control'
Files that had been uploaded to the digitalmarketplace-dev-uploads
bucket didn't have permissions that meant they were accessible from the
console.

The issue is that both buckets and objects have their own different
access controls lists.

The digitalmarketplace-dev-uploads bucket is part of the dev account.
When running the apps locally boto3 will use your default credentials as
specified in ~/.aws/credentials. If set up correctly, these will be your
credentials for the main account (this is the account that your user is
set up on). This means that any time you upload a file, it's coming from
the main account. In S3, when you look at the permissions for an object,
the name you see (suppliers or development-aws in our case) is the
username of the root user for that account. I think that the main
account root user name is suppliers, as it was set up with the email
address supplier@digitalmarketplace.service.gov.uk.

If a file is uploaded to an s3 bucket by an account that isn't the
bucket owner, that object does not inherit the permissions of the
bucket. That's to say that the bucket owner doesn't automatically get
all the permissions they would had they uploaded the object themselves.
The account that uploaded the object needs to apply an access control
list that allows the bucket owner increased permissions.

AWS has some "canned ACLs" which can be applied to buckets/objects
(https://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html).
Currently, when uploading documents we use the private ACL which gives
the object owner full permissions but the bucket owner none
(https://github.com/alphagov/digitalmarketplace-supplier-frontend/blob/master/app/main/views/frameworks.py#L993).
We probably shouldn't do this. We should instead use
bucket-owner-full-control which gives both the bucket and object full
control over the object.