Add default empty option to country picker
If no country data is passed in, we need to show an empty option instead
of the first in the list.

Also fix bug with country picker and move javascripts:

The country picker would send form data for the last country selected.
If you selected a country, and the erased everything in the box, when
you hit submit it would still send data from that country, rather than
the validation failing.

This is because of the way the javascript works and that a select box
can't usually be empty once selected. A small bit of custom JS here will
clear the value of the submitted data if the length of the inputed
country is less than 2. This reflects the way the autocomplete only
kicks in for inputs of two characters and above.

It also moves the JS to the template rather than the base page so it
isn't loaded unnecessarily on every page. This may not be the best place
for it, perhaps it should be moved to the javascripts dir in the assets.
Someone better at JS will know the answer.