Apply validation styling manually for country picker field
Since the country picker field is dynamically managed by
its scripts that change the class names depending on state
our standard validation styling wrapper, which uses
class, does not work on this field. Hence, we had to apply
this styling manually by targetting the field id.

Part of this work was to target country field with JS:

There is some really odd behaviour exhibited by the validation masthead
links, when linking to the country picker input field when JavaScript is
enabled.

The link can't seem to find the input field. However, by moving the
anchor element to elsewhere in the html document (via chrome devtools)
for example, the link can find the element. We can't find a reason for
this and believe it has something to do with the JavaScript being
applied to the page by the country picker.

A workaround for this is to use JavaScript to power the link. As the
form only has one field, we can change the use of the `field_name`
variable to a hardcoded JS thing that targets the field id.

Apply conditional styling to autocomplete hint

The hint text in the country picker autocomplete needs slightly
different styling depending on if the field has a validation error or
not. Without the extra pixels for the border, the text becomes out of
line with the users input.

Extra padding is also applied, for the same reasons. This custom styling
is added to our own application css, and applied after the country
pickers own css to ensure that the custom styling is applied.