Load list of countries in config
Previously we were opening the file with the list of countries every
time a user hit the GET or POST routes associated with the country
picker. It would be better to just load this file once and keep it in
memory.

Loading in the config allows us to easily use this list of countries in
several places. It will be useful when replaying country information to
the user.

Defining the `load_countries` method in the config may not be the best
place for it, however it can be moved if necessary.

Also fix style to please flake8