Load list of countries in helpers
Previously we were opening the file with the list of countries every
time a user hit the GET or POST routes associated with the country
picker. It would be better to just load this file once and keep it in
memory.

Adding a helper that will load the file once when the app starts and
makes it available as a constant that can be imported