Update the company details page to use new country names
Now we can convert the new country codes to country names, we can use
them in the details page.

We need to support the old style country code ('gb'), that we're
currently using, until we've updated all the data in the db to use the
new style. Once this has been released and the data migrated, we can
drop support for it.