Move supplier specific macros to tookit
There were some macros in the supplier frontend which are useful so we
want to keep them. This means they should be in a shared place.

The macros wrap the form fields defined in the tookit. The toolkit
already has some macros for doing this, but they're designed for
situations where we have a question object from the content loader. It's
useful to have some that don't.