Refactor inline JavaScript to separate file
This required a few things to happen.

As this JS is very page specific there's no need to load it on every
page of the app. As such, rather than compiling it with application.js
it's being loaded separately.

A new block, `page_scripts`, has been created on the base page as a place
to put links to page specific javascripts. This allows them to be
included in the `body_end` block without having to call `super`.

A new copy factory has been added to the gulpfile to make sure we have
access to the script in the static directory.