Handle suppliers with no registrationCountry key
When a suppliers are serialized by the api any `None` value keys are
dropped before being returned in the response.

The `registrationCountry` field in the suppliers table is a nullable
field. Whilst in practice in production we don't have any null values in
this field (it's either an empty string or 'gb' at the time of writing),
we should cover this case just in case.

This issue was picked up by the functional test on preview. The
suppliers generated by the functional tests have null values for
`registrationCountry`. This may also need looking at.