Add validation mastheads
When https://trello.com/c/eB9u3xe7/40 was getting QA'd, it was
highlighted that we hadn't added a validation masthead at the top of the
page to highlight that something was wrong. This has now been added and
retrofitted to some other single-question-pages that were missing the
same, so that we're more consistent.