Add "one-time" page to add registered company name
Suppliers need to add their registered company name, but once it has been set they need to contact support to get it changed.

This adds a new page with a form to set registered company name, along with a reusable template for future "one-time" pages that explains to suppliers that they need to contact support in order to change certain details.