Add VatNumberForm
We need only validate the VAT number entered if the user has stated that
they are registered. We can't use the `Optional` validator for this in
case the user has started typing something in the input field and then
changed their mind.

The regex used for the VAT number covers 9 and 12 digit VAT codes, both
with the optional 'GB' prefix. It also supports the lesser used but
still valid GD and HA vat numbers which are used for government
departments and health authorities. Again, with the optional 'GB'
prefix.