Remove JavaScript link from validation masthead
There was an issue with the usual links not working so we used
JavaScript instead. I've worked out what the problem was.

The links are actually powered by some other JavaScript from the
frontend toolkit. They look for an element with an ID that matches the
field name, and then target the first input or text area _within_ that.
So if the input itself has the id, or the id isn't used, the link won't
work.