Create helper for masthead form_errors
It allows single or multiple forms to be passed in and their validation
errors to be processed into one list. This can be fed to templates as a
variable rather than doing the logic in the template.

The tests create a form class instead of a mock, as the mock was a pain
to set up.