Tweak validation of company number form
There was an issue where previously entered "bad" data in one company number field could prevent successful submission of valid data in the other company field, even though the "bad" field wasn't visible on the page when the form was submitted.

This fixes the validator to clear any data in fields that were not visible when the form was submitted.  (Or, in the case of no-Javascript users, clearing data from fields that were not associated with the selected "Yes/No" radio button.