Update button content for copying service
When copying a draft service, after clicking the 'copy' button, the
first question you saw had a 'Save and return' button. Clicking that
though took you to the service summary, instead of the service list,
which is inconsistent. 'Save and continue' seemed a better choice.

This required the `force_continue_button` argument to be passed in the
redirect from the copy view and accepted in the `edit_service_submission`
view.

Support for passing the `return_to_summary` argument to the
`edit_service_submission` view has also been removed, as it was not
being used anywhere in the app.