Refactor application started emails
 ## Summary
Refactors the emails which get sent out when a supplier user starts an
application with a framework. Instead of having a template for each
framework iteration, we can have one for each framework family and
inject the required dynamic content using the content-loader. This
should mean we don't need to update this template every time we run a
new framework.

In theory it might be possible to refactor this further and have a
single 'application started' template for all frameworks - but it seems
reasonable to allow differences in the email between frameworks in case
guidance needs to be materially different.