Remove CSRF token from form data
 ## Summary
An update to FlaskWTF now causes `csrf_token` to appear in WTForm's
`form.data`. Our API doesn't like this so we need to strip it out when
we use the entire `form.data` as opposed to `form.<question>.data`. A
small utility in dm-utils will allow us to do this easily if end up
needing it in other apps.